# Implementation-only import checking for conformances