--- a/lib/Sema/TypeCheckAccess.cpp
+++ b/lib/Sema/TypeCheckAccess.cpp
@@ -22,6 +22,7 @@
 #include "swift/AST/ExistentialLayout.h"
 #include "swift/AST/Pattern.h"
 #include "swift/AST/ParameterList.h"
+#include "swift/AST/ProtocolConformance.h"
 #include "swift/AST/TypeCheckRequests.h"
 
 using namespace swift;
@@ -1402,13 +1403,17 @@ public:
 
 class ImplementationOnlyImportChecker
     : public DeclVisitor<ImplementationOnlyImportChecker> {
-  using CheckImplementationOnlyCallback =
+  using CheckImplementationOnlyTypeCallback =
       llvm::function_ref<void(const TypeDecl *, const TypeRepr *)>;
+  using CheckImplementationOnlyConformanceCallback =
+      llvm::function_ref<void(const ProtocolConformance *)>;
 
   TypeChecker &TC;
 
-  void checkTypeImpl(Type type, const TypeRepr *typeRepr, const SourceFile *SF,
-                     CheckImplementationOnlyCallback diagnose) {
+  void checkTypeImpl(
+      Type type, const TypeRepr *typeRepr, const SourceFile &SF,
+      CheckImplementationOnlyTypeCallback diagnoseType,
+      CheckImplementationOnlyConformanceCallback diagnoseConformance) {
     // Don't bother checking errors.
     if (type && type->hasError())
       return;
@@ -1421,10 +1426,10 @@ class ImplementationOnlyImportChecker
       const_cast<TypeRepr *>(typeRepr)->walk(TypeReprIdentFinder(
           [&](const ComponentIdentTypeRepr *component) {
         ModuleDecl *M = component->getBoundDecl()->getModuleContext();
-        if (!SF->isImportedImplementationOnly(M))
+        if (!SF.isImportedImplementationOnly(M))
           return true;
 
-        diagnose(component->getBoundDecl(), component);
+        diagnoseType(component->getBoundDecl(), component);
         foundAnyIssues = true;
         // We still continue even in the diagnostic case to report multiple
         // violations.
@@ -1432,32 +1437,93 @@ class ImplementationOnlyImportChecker
       }));
     }
 
-    // FIXME: Are there places where we can be sure the TypeRepr fully accounts
-    // for the type, and therefore we could skip this extra work?
-    if (!foundAnyIssues && type) {
-      type.walk(SimpleTypeDeclFinder([&](const TypeDecl *typeDecl) {
-        ModuleDecl *M = typeDecl->getModuleContext();
-        if (!SF->isImportedImplementationOnly(M))
-          return TypeWalker::Action::Continue;
+    // Note that if we have a type, we can't skip checking it even if the
+    // TypeRepr is okay, because that's how we check what conformances are
+    // being used.
+    //
+    // We still don't want to do this if we found issues with the TypeRepr,
+    // though, because that would result in some issues being reported twice.
+    if (foundAnyIssues || !type)
+      return;
 
-        diagnose(typeDecl, /*typeRepr*/nullptr);
-        // We still continue even in the diagnostic case to report multiple
-        // violations.
-        return TypeWalker::Action::Continue;
-      }));
-    }
+    class ProblematicTypeFinder : public TypeDeclFinder {
+      const SourceFile &SF;
+      CheckImplementationOnlyTypeCallback diagnoseType;
+      CheckImplementationOnlyConformanceCallback diagnoseConformance;
+    public:
+      ProblematicTypeFinder(
+          const SourceFile &SF,
+          CheckImplementationOnlyTypeCallback diagnoseType,
+          CheckImplementationOnlyConformanceCallback diagnoseConformance)
+        : SF(SF), diagnoseType(diagnoseType),
+          diagnoseConformance(diagnoseConformance) {}
+
+      void visitTypeDecl(const TypeDecl *typeDecl) {
+        ModuleDecl *M = typeDecl->getModuleContext();
+        if (!SF.isImportedImplementationOnly(M))
+          return;
+
+        diagnoseType(typeDecl, /*typeRepr*/nullptr);
+      }
+
+      void visitSubstitutionMap(const SubstitutionMap &subs) {
+        for (ProtocolConformanceRef conformance : subs.getConformances()) {
+          if (!conformance.isConcrete())
+            continue;
+          const ProtocolConformance *concreteConf = conformance.getConcrete();
+
+          SubstitutionMap subConformanceSubs =
+              concreteConf->getSubstitutions(SF.getParentModule());
+          visitSubstitutionMap(subConformanceSubs);
+
+          const RootProtocolConformance *rootConf =
+              concreteConf->getRootConformance();
+          ModuleDecl *M = rootConf->getDeclContext()->getParentModule();
+          if (!SF.isImportedImplementationOnly(M))
+            continue;
+          diagnoseConformance(rootConf);
+        }
+      }
+
+      Action visitNominalType(NominalType *ty) override {
+        visitTypeDecl(ty->getDecl());
+        return Action::Continue;
+      }
+
+      Action visitBoundGenericType(BoundGenericType *ty) override {
+        visitTypeDecl(ty->getDecl());
+        SubstitutionMap subs =
+            ty->getContextSubstitutionMap(SF.getParentModule(), ty->getDecl());
+        visitSubstitutionMap(subs);
+        return Action::Continue;
+      }
+
+      Action visitTypeAliasType(TypeAliasType *ty) override {
+        visitTypeDecl(ty->getDecl());
+        visitSubstitutionMap(ty->getSubstitutionMap());
+        return Action::Continue;
+      }
+    };
+
+    type.walk(ProblematicTypeFinder(SF, diagnoseType, diagnoseConformance));
   }
 
-  void checkType(Type type, const TypeRepr *typeRepr, const Decl *context,
-                 CheckImplementationOnlyCallback diagnose) {
+  void checkType(
+      Type type, const TypeRepr *typeRepr, const Decl *context,
+      CheckImplementationOnlyTypeCallback diagnoseType,
+      CheckImplementationOnlyConformanceCallback diagnoseConformance) {
     auto *SF = context->getDeclContext()->getParentSourceFile();
     assert(SF && "checking a non-source declaration?");
-    return checkTypeImpl(type, typeRepr, SF, diagnose);
+    return checkTypeImpl(type, typeRepr, *SF, diagnoseType,
+                         diagnoseConformance);
   }
 
-  void checkType(const TypeLoc &TL, const Decl *context,
-                 CheckImplementationOnlyCallback diagnose) {
-    checkType(TL.getType(), TL.getTypeRepr(), context, diagnose);
+  void checkType(
+      const TypeLoc &TL, const Decl *context,
+      CheckImplementationOnlyTypeCallback diagnoseType,
+      CheckImplementationOnlyConformanceCallback diagnoseConformance) {
+    checkType(TL.getType(), TL.getTypeRepr(), context, diagnoseType,
+              diagnoseConformance);
   }
 
   void checkGenericParams(const GenericParamList *params,
@@ -1470,14 +1536,15 @@ class ImplementationOnlyImportChecker
         continue;
       assert(param->getInherited().size() == 1);
       checkType(param->getInherited().front(), owner,
-                getDiagnoseCallback(owner));
+                getDiagnoseCallback(owner), getDiagnoseCallback(owner));
     }
 
     forAllRequirementTypes(WhereClauseOwner(
                              owner->getInnermostDeclContext(),
                              const_cast<GenericParamList *>(params)),
                            [&](Type type, TypeRepr *typeRepr) {
-      checkType(type, typeRepr, owner, getDiagnoseCallback(owner));
+      checkType(type, typeRepr, owner, getDiagnoseCallback(owner),
+                getDiagnoseCallback(owner));
     });
   }
 
@@ -1494,11 +1561,24 @@ class ImplementationOnlyImportChecker
                               offendingType->getFullName(), M->getName());
       highlightOffendingType(TC, diag, complainRepr);
     }
+
+    void operator()(const ProtocolConformance *offendingConformance) {
+      ModuleDecl *M = offendingConformance->getDeclContext()->getParentModule();
+      TC.diagnose(D, diag::conformance_from_implementation_only_module,
+                  offendingConformance->getType(),
+                  offendingConformance->getProtocol()->getFullName(),
+                  M->getName());
+    }
   };
 
-  static_assert(std::is_convertible<DiagnoseGenerically,
-                                    CheckImplementationOnlyCallback>::value,
-                "DiagnoseGenerically has wrong call signature");
+  static_assert(
+      std::is_convertible<DiagnoseGenerically,
+                          CheckImplementationOnlyTypeCallback>::value,
+      "DiagnoseGenerically has wrong call signature");
+  static_assert(
+      std::is_convertible<DiagnoseGenerically,
+                          CheckImplementationOnlyConformanceCallback>::value,
+      "DiagnoseGenerically has wrong call signature for conformance diags");
 
   DiagnoseGenerically getDiagnoseCallback(const Decl *D) {
     return DiagnoseGenerically(TC, D);
@@ -1544,7 +1624,7 @@ public:
 
     if (auto *extension = dyn_cast<ExtensionDecl>(D->getDeclContext())) {
       checkType(extension->getExtendedTypeLoc(), extension,
-                getDiagnoseCallback(extension));
+                getDiagnoseCallback(extension), getDiagnoseCallback(extension));
       checkConstrainedExtensionRequirements(extension);
     }
   }
@@ -1593,7 +1673,7 @@ public:
       return;
 
     checkType(theVar->getInterfaceType(), /*typeRepr*/nullptr, theVar,
-              getDiagnoseCallback(theVar));
+              getDiagnoseCallback(theVar), getDiagnoseCallback(theVar));
   }
 
   /// \see visitPatternBindingDecl
@@ -1612,7 +1692,8 @@ public:
     if (shouldSkipChecking(anyVar))
       return;
 
-    checkType(TP->getTypeLoc(), anyVar, getDiagnoseCallback(anyVar));
+    checkType(TP->getTypeLoc(), anyVar, getDiagnoseCallback(anyVar),
+              getDiagnoseCallback(anyVar));
   }
 
   void visitPatternBindingDecl(PatternBindingDecl *PBD) {
@@ -1635,21 +1716,24 @@ public:
 
   void visitTypeAliasDecl(TypeAliasDecl *TAD) {
     checkGenericParams(TAD->getGenericParams(), TAD);
-    checkType(TAD->getUnderlyingTypeLoc(), TAD, getDiagnoseCallback(TAD));
+    checkType(TAD->getUnderlyingTypeLoc(), TAD, getDiagnoseCallback(TAD),
+              getDiagnoseCallback(TAD));
   }
 
   void visitAssociatedTypeDecl(AssociatedTypeDecl *assocType) {
     llvm::for_each(assocType->getInherited(),
                    [&](TypeLoc requirement) {
-      checkType(requirement, assocType, getDiagnoseCallback(assocType));
+      checkType(requirement, assocType, getDiagnoseCallback(assocType),
+                getDiagnoseCallback(assocType));
     });
     checkType(assocType->getDefaultDefinitionLoc(), assocType,
-              getDiagnoseCallback(assocType));
+              getDiagnoseCallback(assocType), getDiagnoseCallback(assocType));
 
     if (assocType->getTrailingWhereClause()) {
       forAllRequirementTypes(assocType,
                              [&](Type type, TypeRepr *typeRepr) {
-        checkType(type, typeRepr, assocType, getDiagnoseCallback(assocType));
+        checkType(type, typeRepr, assocType, getDiagnoseCallback(assocType),
+                  getDiagnoseCallback(assocType));
       });
     }
   }
@@ -1659,19 +1743,22 @@ public:
 
     llvm::for_each(nominal->getInherited(),
                    [&](TypeLoc nextInherited) {
-      checkType(nextInherited, nominal, getDiagnoseCallback(nominal));
+      checkType(nextInherited, nominal, getDiagnoseCallback(nominal),
+                getDiagnoseCallback(nominal));
     });
   }
 
   void visitProtocolDecl(ProtocolDecl *proto) {
     llvm::for_each(proto->getInherited(),
                   [&](TypeLoc requirement) {
-      checkType(requirement, proto, getDiagnoseCallback(proto));
+      checkType(requirement, proto, getDiagnoseCallback(proto),
+                getDiagnoseCallback(proto));
     });
 
     if (proto->getTrailingWhereClause()) {
       forAllRequirementTypes(proto, [&](Type type, TypeRepr *typeRepr) {
-        checkType(type, typeRepr, proto, getDiagnoseCallback(proto));
+        checkType(type, typeRepr, proto, getDiagnoseCallback(proto),
+                  getDiagnoseCallback(proto));
       });
     }
   }
@@ -1679,35 +1766,42 @@ public:
   void visitSubscriptDecl(SubscriptDecl *SD) {
     checkGenericParams(SD->getGenericParams(), SD);
 
-    for (auto &P : *SD->getIndices())
-      checkType(P->getTypeLoc(), SD, getDiagnoseCallback(SD));
-    checkType(SD->getElementTypeLoc(), SD, getDiagnoseCallback(SD));
+    for (auto &P : *SD->getIndices()) {
+      checkType(P->getTypeLoc(), SD, getDiagnoseCallback(SD),
+                getDiagnoseCallback(SD));
+    }
+    checkType(SD->getElementTypeLoc(), SD, getDiagnoseCallback(SD),
+              getDiagnoseCallback(SD));
   }
 
   void visitAbstractFunctionDecl(AbstractFunctionDecl *fn) {
     checkGenericParams(fn->getGenericParams(), fn);
 
     for (auto *P : *fn->getParameters())
-      checkType(P->getTypeLoc(), fn, getDiagnoseCallback(fn));
+      checkType(P->getTypeLoc(), fn, getDiagnoseCallback(fn),
+                getDiagnoseCallback(fn));
   }
 
   void visitFuncDecl(FuncDecl *FD) {
     visitAbstractFunctionDecl(FD);
-    checkType(FD->getBodyResultTypeLoc(), FD, getDiagnoseCallback(FD));
+    checkType(FD->getBodyResultTypeLoc(), FD, getDiagnoseCallback(FD),
+              getDiagnoseCallback(FD));
   }
 
   void visitEnumElementDecl(EnumElementDecl *EED) {
     if (!EED->hasAssociatedValues())
       return;
     for (auto &P : *EED->getParameterList())
-      checkType(P->getTypeLoc(), EED, getDiagnoseCallback(EED));
+      checkType(P->getTypeLoc(), EED, getDiagnoseCallback(EED),
+                getDiagnoseCallback(EED));
   }
 
   void checkConstrainedExtensionRequirements(ExtensionDecl *ED) {
     if (!ED->getTrailingWhereClause())
       return;
     forAllRequirementTypes(ED, [&](Type type, TypeRepr *typeRepr) {
-      checkType(type, typeRepr, ED, getDiagnoseCallback(ED));
+      checkType(type, typeRepr, ED, getDiagnoseCallback(ED),
+                getDiagnoseCallback(ED));
     });
   }
 
@@ -1719,7 +1813,8 @@ public:
     // but just hide that from interfaces.
     llvm::for_each(ED->getInherited(),
                    [&](TypeLoc nextInherited) {
-      checkType(nextInherited, ED, getDiagnoseCallback(ED));
+      checkType(nextInherited, ED, getDiagnoseCallback(ED),
+                getDiagnoseCallback(ED));
     });
 
     if (!ED->getInherited().empty())