--- a/test/Sema/Inputs/implementation-only-import-in-decls-helper.swift
+++ b/test/Sema/Inputs/implementation-only-import-in-decls-helper.swift
@@ -1,3 +1,15 @@
+import NormalLibrary
+
+extension NormalStruct: NormalProto {
+  public typealias Assoc = Int
+}
+extension GenericStruct: NormalProto {
+  public typealias Assoc = Int
+}
+extension NormalClass: NormalProto {
+  public typealias Assoc = Int
+}
+
 public struct BadStruct {}
 public protocol BadProto {}
 open class BadClass {}