--- a/test/Sema/Inputs/implementation-only-import-in-decls-public-helper.swift
+++ b/test/Sema/Inputs/implementation-only-import-in-decls-public-helper.swift
@@ -1,0 +1,6 @@
+public struct NormalStruct {}
+public struct GenericStruct<T> {}
+public protocol NormalProto {
+  associatedtype Assoc
+}
+open class NormalClass {}