… lib/Sema/TypeCheckAvailability.cpp
… lib/Sema/TypeCheckProtocol.cpp
… test/Sema/implementation-only-import-inlinable-conformances.swift
… test/Sema/implementation-only-import-library-evolution.swift