@jrose-apple 2019-04-09
Okay, I think this is ready. I removed a bunch of duplication from the last commit and now things aren't *perfect* but they're at least not *terribly bad.*

What test cases am I missing?

Leaving marked as a draft because it still slurps in #23912, so I have to do one more rebase.

I'll do a follow-up PR that harmonizes John's and my terminology, since right now half of the stuff says "exportability" and the other half says "implementation-only" or "implementationOnly"; and another one that renames AccessScopeChecker.h to TypeDeclFinder.h (and pulls the access control walking out) as requested by Slava.

@swift-ci Please test