@jrose-apple 2019-04-09
@swift-ci Please test compiler performance