@jrose-apple 2019-04-10
@swift-ci Please test compiler performance