@jrose-apple 2019-04-10
<details>
<summary>Compiler perf results</summary>

# Summary for master full

**Unexpected test results, excluded stats for ProcedureKit, Tagged, Wordy, Deferred**

**Regressions found (see below)**

- [Debug-batch](#debug-batch-72696f4f12fd8fcf)
    - [debug-batch brief](#debug-batch-brief-72696f4f12fd8fcf)
    - [debug-batch detailed](#debug-batch-detailed-72696f4f12fd8fcf)
- [Release](#release-72696f4f12fd8fcf)
    - [release brief](#release-brief-72696f4f12fd8fcf)
    - [release detailed](#release-detailed-72696f4f12fd8fcf)



<a name="debug-batch-72696f4f12fd8fcf">

# Debug-batch

<a name="debug-batch-brief-72696f4f12fd8fcf">

## debug-batch brief
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (2)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 29,895,438,266,079 | 29,525,082,392,802 | -370,355,873,277 | -1.24% :white_check_mark:
time.swift-driver.wall | 2804.4s | 2749.5s | -54.8s | -1.95% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
LLVM.NumLLVMBytesOutput | 1,114,053,216 | 1,114,053,366 | 150 | 0.0%
</details>

<a name="debug-batch-detailed-72696f4f12fd8fcf">

## debug-batch detailed
<details>
<summary>Regressed (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.NumSourceLinesPerSecond | 2,251,754 | 2,297,279 | 45,525 | 2.02% :no_entry:
</details>
<details>
<summary>Improved (6)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 29,895,438,266,079 | 29,525,082,392,802 | -370,355,873,277 | -1.24% :white_check_mark:
Sema.NumConstraintScopes | 18,011,511 | 17,540,127 | -471,384 | -2.62% :white_check_mark:
Sema.NumConstraintsConsideredForEdgeContraction | 45,554,258 | 43,388,452 | -2,165,806 | -4.75% :white_check_mark:
Sema.NumLeafScopes | 12,083,326 | 11,784,684 | -298,642 | -2.47% :white_check_mark:
Sema.OverriddenDeclsRequest | 8,060,968 | 7,959,067 | -101,901 | -1.26% :white_check_mark:
Sema.USRGenerationRequest | 13,133,931 | 12,922,649 | -211,282 | -1.61% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (93)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.NumASTBytesAllocated | 71,639,257,132 | 71,029,843,000 | -609,414,132 | -0.85%
AST.NumDecls | 92,845 | 92,845 | 0 | 0.0%
AST.NumDependencies | 212,128 | 212,130 | 2 | 0.0%
AST.NumImportedExternalDefinitions | 1,199,076 | 1,199,076 | 0 | 0.0%
AST.NumInfixOperators | 34,255 | 34,255 | 0 | 0.0%
AST.NumLinkLibraries | 0 | 0 | 0 | 0.0%
AST.NumLoadedModules | 253,056 | 253,056 | 0 | 0.0%
AST.NumLocalTypeDecls | 123 | 123 | 0 | 0.0%
AST.NumObjCMethods | 15,429 | 15,429 | 0 | 0.0%
AST.NumPostfixOperators | 18 | 18 | 0 | 0.0%
AST.NumPrecedenceGroups | 16,924 | 16,924 | 0 | 0.0%
AST.NumPrefixOperators | 85 | 85 | 0 | 0.0%
AST.NumReferencedDynamicNames | 122 | 122 | 0 | 0.0%
AST.NumReferencedMemberNames | 4,042,754 | 4,042,754 | 0 | 0.0%
AST.NumReferencedTopLevelNames | 313,132 | 313,132 | 0 | 0.0%
AST.NumSourceBuffers | 387,344 | 387,344 | 0 | 0.0%
AST.NumSourceLines | 2,973,052 | 2,973,052 | 0 | 0.0%
AST.NumTotalClangImportedEntities | 4,579,567 | 4,569,384 | -10,183 | -0.22%
AST.NumUsedConformances | 260,546 | 260,546 | 0 | 0.0%
Driver.ChildrenMaxRSS | 110,494,564,352 | 109,887,848,448 | -606,715,904 | -0.55%
Driver.DriverDepCascadingDynamic | 0 | 0 | 0 | 0.0%
Driver.DriverDepCascadingExternal | 0 | 0 | 0 | 0.0%
Driver.DriverDepCascadingMember | 0 | 0 | 0 | 0.0%
Driver.DriverDepCascadingNominal | 0 | 0 | 0 | 0.0%
Driver.DriverDepCascadingTopLevel | 0 | 0 | 0 | 0.0%
Driver.DriverDepDynamic | 0 | 0 | 0 | 0.0%
Driver.DriverDepExternal | 0 | 0 | 0 | 0.0%
Driver.DriverDepMember | 0 | 0 | 0 | 0.0%
Driver.DriverDepNominal | 0 | 0 | 0 | 0.0%
Driver.DriverDepTopLevel | 0 | 0 | 0 | 0.0%
Driver.NumDriverJobsRun | 18,540 | 18,540 | 0 | 0.0%
Driver.NumDriverJobsSkipped | 0 | 0 | 0 | 0.0%
Driver.NumDriverPipePolls | 199,666 | 201,467 | 1,801 | 0.9%
Driver.NumDriverPipeReads | 222,529 | 224,543 | 2,014 | 0.91%
Driver.NumProcessFailures | 0 | 0 | 0 | 0.0%
Frontend.MaxMallocUsage | 729,673,511,696 | 725,924,788,872 | -3,748,722,824 | -0.51%
Frontend.NumProcessFailures | 0 | 0 | 0 | 0.0%
IRModule.NumIRAliases | 115,668 | 115,668 | 0 | 0.0%
IRModule.NumIRBasicBlocks | 4,387,012 | 4,387,012 | 0 | 0.0%
IRModule.NumIRComdatSymbols | 0 | 0 | 0 | 0.0%
IRModule.NumIRFunctions | 2,083,997 | 2,083,997 | 0 | 0.0%
IRModule.NumIRGlobals | 2,104,798 | 2,104,798 | 0 | 0.0%
IRModule.NumIRIFuncs | 0 | 0 | 0 | 0.0%
IRModule.NumIRInsts | 54,813,513 | 54,813,513 | 0 | 0.0%
IRModule.NumIRNamedMetaData | 90,002 | 90,002 | 0 | 0.0%
IRModule.NumIRValueSymbols | 3,787,221 | 3,787,221 | 0 | 0.0%
LLVM.NumLLVMBytesOutput | 1,114,053,216 | 1,114,053,366 | 150 | 0.0%
Parse.NumFunctionsParsed | 168,408 | 168,408 | 0 | 0.0%
Parse.NumIterableDeclContextParsed | 1,163,266 | 1,163,266 | 0 | 0.0%
SILModule.NumSILGenDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenFunctions | 1,071,771 | 1,071,771 | 0 | 0.0%
SILModule.NumSILGenGlobalVariables | 40,103 | 40,103 | 0 | 0.0%
SILModule.NumSILGenVtables | 11,745 | 11,745 | 0 | 0.0%
SILModule.NumSILGenWitnessTables | 45,902 | 45,902 | 0 | 0.0%
SILModule.NumSILOptDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILOptFunctions | 1,502,340 | 1,502,340 | 0 | 0.0%
SILModule.NumSILOptGlobalVariables | 40,984 | 40,984 | 0 | 0.0%
SILModule.NumSILOptVtables | 19,254 | 19,254 | 0 | 0.0%
SILModule.NumSILOptWitnessTables | 99,775 | 99,775 | 0 | 0.0%
Sema.AccessLevelRequest | 2,768,880 | 2,766,417 | -2,463 | -0.09%
Sema.CustomAttrNominalRequest | 0 | 0 | 0 | 0.0%
Sema.DefaultAndMaxAccessLevelRequest | 62,984 | 62,976 | -8 | -0.01%
Sema.DefaultTypeRequest | 370,154 | 370,154 | 0 | 0.0%
Sema.EnumRawTypeRequest | 20,134 | 20,134 | 0 | 0.0%
Sema.ExtendedNominalRequest | 4,060,275 | 4,052,009 | -8,266 | -0.2%
Sema.InheritedDeclsReferencedRequest | 4,751,981 | 4,735,432 | -16,549 | -0.35%
Sema.InheritedTypeRequest | 646,215 | 646,286 | 71 | 0.01%
Sema.IsDynamicRequest | 2,134,078 | 2,134,078 | 0 | 0.0%
Sema.IsObjCRequest | 1,874,699 | 1,873,868 | -831 | -0.04%
Sema.MangleLocalTypeDeclRequest | 246 | 246 | 0 | 0.0%
Sema.NamedLazyMemberLoadFailureCount | 23,284 | 23,232 | -52 | -0.22%
Sema.NamedLazyMemberLoadSuccessCount | 19,496,042 | 19,498,938 | 2,896 | 0.01%
Sema.NominalTypeLookupDirectCount | 33,676,983 | 33,558,979 | -118,004 | -0.35%
Sema.NumConformancesDeserialized | 6,905,659 | 6,870,966 | -34,693 | -0.5%
Sema.NumDeclsDeserialized | 51,165,120 | 50,869,830 | -295,290 | -0.58%
Sema.NumDeclsFinalized | 1,937,417 | 1,937,417 | 0 | 0.0%
Sema.NumDeclsTypechecked | 1,003,408 | 1,003,408 | 0 | 0.0%
Sema.NumDeclsValidated | 2,350,955 | 2,350,955 | 0 | 0.0%
Sema.NumFunctionsTypechecked | 1,063,532 | 1,063,532 | 0 | 0.0%
Sema.NumGenericSignatureBuilders | 1,289,802 | 1,286,472 | -3,330 | -0.26%
Sema.NumLazyGenericEnvironments | 10,308,768 | 10,255,248 | -53,520 | -0.52%
Sema.NumLazyGenericEnvironmentsLoaded | 222,751 | 222,779 | 28 | 0.01%
Sema.NumLazyIterableDeclContexts | 7,153,492 | 7,142,149 | -11,343 | -0.16%
Sema.NumTypesDeserialized | 17,241,293 | 17,177,046 | -64,247 | -0.37%
Sema.NumTypesValidated | 1,795,698 | 1,795,698 | 0 | 0.0%
Sema.NumUnloadedLazyIterableDeclContexts | 4,699,067 | 4,702,954 | 3,887 | 0.08%
Sema.RequirementRequest | 71,867 | 71,867 | 0 | 0.0%
Sema.SelfBoundsFromWhereClauseRequest | 6,815,966 | 6,792,516 | -23,450 | -0.34%
Sema.SetterAccessLevelRequest | 157,241 | 157,241 | 0 | 0.0%
Sema.SuperclassDeclRequest | 86,208 | 86,246 | 38 | 0.04%
Sema.SuperclassTypeRequest | 37,133 | 37,133 | 0 | 0.0%
Sema.TypeDeclsFromWhereClauseRequest | 33,250 | 33,242 | -8 | -0.02%
Sema.UnderlyingTypeDeclsReferencedRequest | 185,120 | 184,873 | -247 | -0.13%
</details>

<a name="release-72696f4f12fd8fcf">

# Release

<a name="release-brief-72696f4f12fd8fcf">

## release brief
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (3)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 29,063,512,320,894 | 29,010,625,697,099 | -52,886,623,795 | -0.18%
LLVM.NumLLVMBytesOutput | 946,121,674 | 946,128,714 | 7,040 | 0.0%
time.swift-driver.wall | 5084.9s | 5085.1s | 203.7ms | 0.0%
</details>

<a name="release-detailed-72696f4f12fd8fcf">

## release detailed
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Sema.NumConstraintScopes | 16,161,421 | 15,689,803 | -471,618 | -2.92% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (22)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.NumImportedExternalDefinitions | 230,987 | 230,987 | 0 | 0.0%
AST.NumLoadedModules | 16,804 | 16,804 | 0 | 0.0%
AST.NumTotalClangImportedEntities | 792,872 | 792,872 | 0 | 0.0%
AST.NumUsedConformances | 261,462 | 261,462 | 0 | 0.0%
IRModule.NumIRBasicBlocks | 3,844,580 | 3,844,580 | 0 | 0.0%
IRModule.NumIRFunctions | 1,754,103 | 1,754,103 | 0 | 0.0%
IRModule.NumIRGlobals | 1,856,775 | 1,856,775 | 0 | 0.0%
IRModule.NumIRInsts | 34,649,265 | 34,649,265 | 0 | 0.0%
IRModule.NumIRValueSymbols | 3,367,396 | 3,367,396 | 0 | 0.0%
LLVM.NumLLVMBytesOutput | 946,121,674 | 946,128,714 | 7,040 | 0.0%
SILModule.NumSILGenFunctions | 743,986 | 743,986 | 0 | 0.0%
SILModule.NumSILOptFunctions | 1,001,526 | 1,001,526 | 0 | 0.0%
Sema.NumConformancesDeserialized | 2,334,731 | 2,334,655 | -76 | -0.0%
Sema.NumDeclsDeserialized | 6,257,924 | 6,255,677 | -2,247 | -0.04%
Sema.NumDeclsValidated | 1,208,498 | 1,208,498 | 0 | 0.0%
Sema.NumFunctionsTypechecked | 479,126 | 479,126 | 0 | 0.0%
Sema.NumGenericSignatureBuilders | 210,562 | 210,376 | -186 | -0.09%
Sema.NumLazyGenericEnvironments | 1,290,079 | 1,289,512 | -567 | -0.04%
Sema.NumLazyGenericEnvironmentsLoaded | 22,436 | 22,436 | 0 | 0.0%
Sema.NumLazyIterableDeclContexts | 795,511 | 795,504 | -7 | -0.0%
Sema.NumTypesDeserialized | 3,330,517 | 3,329,434 | -1,083 | -0.03%
Sema.NumTypesValidated | 734,602 | 734,602 | 0 | 0.0%
</details>
</details>

@brentdax, any idea [why it's failing to post the comment](https://ci.swift.org/job/swift-PR-compiler-performance-macOS/372/)? It says the build failed but I can't see how it failed, and at the same time the results look strange again.